[UPDATE]resetlogin add phone and username

+++ Laravel/app/Api/v1/Controllers/ResetLoginController.php
@@ -15,6 +15,7 @@ use App\Api\v1\Utilities\ErrorCodeUtility;
 
 use Auth;
 use Mail;
+use Twilio;
 
 use Carbon\Carbon;
 // use App\Http\Requests;
@@ -35,7 +36,9 @@ class ResetLoginController extends Controller
         }
         
         $validator = Validator::make($input, [
-            'email' => 'required|max:50|email'
+            'email' => 'required_without:phone|max:50|email',
+            'phone' => 'required_without:email|max:20|regex:/^\([0-9]+\)[0-9]+$/',
+            'user_name' => 'required|regex:/^[a-zA-Z0-9]*[_\-.]?[a-zA-Z0-9]*$/|min:3|max:20',
         ]);
         
         if($validator->fails())
@@ -43,7 +46,17 @@ class ResetLoginController extends Controller
     		throw new StoreResourceFailedException('Could not verify.',$validator->errors());
     	}
         
-        $user = Users::where('email','=',$input['email'])->get();
+        $input_key = '';
+        $input_value = '';
+        if($this->request->has('email')){
+            $input_key = 'email';
+            $input_value = $input['email'];
+        }else{
+            $input_key = 'phone';
+            $input_value = $input['phone'];
+        }
+
+        $user = Users::where($input_key, '=', $input_value)->where('user_name', '=', $input['user_name'])->get();
         if( is_null($user) || ( count($user)==0 ) ){
             return response()->json([
                     'message' => 'user not found',
@@ -54,11 +67,16 @@ class ResetLoginController extends Controller
         
         // generate code & store user_id and code to verification table
         $verification_code = (string)rand(111111,999999);
-        $verification = Verifications::where('email','=', $input['email'])->where('type','=', 'resetpassword')->get();
+        $verification = Verifications::where($input_key,'=', $input_value)->where('type','=', 'resetpassword')->get();
         if( is_null($verification) || ( count($verification)==0 ) ){
             $new_verification = new Verifications;
             $new_verification->type = 'resetpassword';
-            $new_verification->email = $input['email'];
+            if( $input_key == 'email' ){
+                $new_verification->email = $input_value;
+            }else{
+                $new_verification->phone = $input_value;
+            }
+            
             $new_verification->code = $verification_code;
             $new_verification->save();
         }else{
@@ -67,7 +85,11 @@ class ResetLoginController extends Controller
                 
                 $new_verification = new Verifications;
                 $new_verification->type = 'resetpassword';
-                $new_verification->email = $input['email'];
+                if( $input_key == 'email' ){
+                    $new_verification->email = $input_value;
+                }else{
+                    $new_verification->phone = $input_value;
+                }
                 $new_verification->code = $verification_code;
                 $new_verification->save();
             }else{
@@ -75,19 +97,25 @@ class ResetLoginController extends Controller
             }
         }
         
-        // send code to this email
-        $email = "Hello!\n\n";
-        $email = $email."This is Fae Support. Below is the code for resetting your password. Please enter the code on your device and proceed to creating your new password. The code is valid for 3 hours or when a new one is issued.\n";
-        $email = $email."\n".$verification_code."\n\n";
-        $email = $email."Have a wonderful day!\n";
-        $email = $email."Fae Support\n\n";
-        $email = $email."*If you did not request this change then please ignore the email.\n";
-        
-        Mail::raw($email, function ($message) {
-            $message->from('support@letsfae.com', 'Fae Support');
 
-            $message->to($this->request->email)->subject('Fae-Reset your password');
-        });
+        if( $input_key == 'email' ){
+            // send code to this email
+            $email = "Hello!\n\n";
+            $email = $email."This is Fae Support. Below is the code for resetting your password. Please enter the code on your device and proceed to creating your new password. The code is valid for 3 hours or when a new one is issued.\n";
+            $email = $email."\n".$verification_code."\n\n";
+            $email = $email."Have a wonderful day!\n";
+            $email = $email."Fae Support\n\n";
+            $email = $email."*If you did not request this change then please ignore the email.\n";
+            
+            Mail::raw($email, function ($message) {
+                $message->from('support@letsfae.com', 'Fae Support');
+
+                $message->to($this->request->email)->subject('Fae-Reset your password');
+            });
+        }else{
+            Twilio::message($this->request->phone, $verification_code);
+        }
+        
 
         return $this->response->created();
         
@@ -101,17 +129,37 @@ class ResetLoginController extends Controller
         }
         
         $validator = Validator::make($input, [
-            'email' => 'required|max:50|email',
+            'email' => 'required_without:phone|max:50|email',
+            'phone' => 'required_without:email|max:20|regex:/^\([0-9]+\)[0-9]+$/',
+            'user_name' => 'required|regex:/^[a-zA-Z0-9]*[_\-.]?[a-zA-Z0-9]*$/|min:3|max:20',
             'code' => 'required|string|max:6'
         ]);
         
-        if($validator->fails())
-    	{
+        if($validator->fails()){
     		throw new StoreResourceFailedException('Could not verify.',$validator->errors());
     	}
-        
+
+        $input_key = '';
+        $input_value = '';
+        if($this->request->has('email')){
+            $input_key = 'email';
+            $input_value = $input['email'];
+        }else{
+            $input_key = 'phone';
+            $input_value = $input['phone'];
+        }
+
+        $user = Users::where($input_key, '=', $input_value)->where('user_name', '=', $input['user_name'])->get();
+        if( is_null($user) || ( count($user)==0 ) ){
+            return response()->json([
+                    'message' => 'user not found',
+                    'error_code' => ErrorCodeUtility::USER_NOT_FOUND,
+                    'status_code' => '404'
+                ], 404);
+        }
+
         // compare user_id and code with data in verification table
-        $verification = Verifications::where('email','=', $input['email'])->where('type','=', 'resetpassword')->get();
+        $verification = Verifications::where($input_key,'=', $input_value)->where('type','=', 'resetpassword')->get();
         if( is_null($verification) || ( count($verification)==0 ) ){
             return response()->json([
                     'message' => 'verification not found',
@@ -147,18 +195,29 @@ class ResetLoginController extends Controller
         }
         
         $validator = Validator::make($input, [
-            'email' => 'required|max:50|email',
+            'email' => 'required_without:phone|max:50|email',
+            'phone' => 'required_without:email|max:20|regex:/^\([0-9]+\)[0-9]+$/',
+            'user_name' => 'required|regex:/^[a-zA-Z0-9]*[_\-.]?[a-zA-Z0-9]*$/|min:3|max:20',
             'password' => 'required|between:8,16',
             'code' => 'required|string|max:6'
         ]);
         
-        if($validator->fails())
-    	{
+        if($validator->fails()) {
     		throw new StoreResourceFailedException('Could not reset.',$validator->errors());
     	}
         
+        $input_key = '';
+        $input_value = '';
+        if($this->request->has('email')){
+            $input_key = 'email';
+            $input_value = $input['email'];
+        }else{
+            $input_key = 'phone';
+            $input_value = $input['phone'];
+        }
+
         // compare user_id and code with data in verification table
-        $verification = Verifications::where('email','=', $input['email'])->where('type','=', 'resetpassword')->get();
+        $verification = Verifications::where($input_key,'=', $input_value)->where('type','=', 'resetpassword')->get();
         if( is_null($verification) || ( count($verification)==0 ) ){
             return response()->json([
                     'message' => 'verification not found',
@@ -174,13 +233,21 @@ class ResetLoginController extends Controller
                         'status_code' => '403'
                     ], 403);
                 }
-                $verification[0]->delete();
                 
-                $user = Users::where('email','=',$input['email'])->get();
+                $user = Users::where($input_key, '=', $input_value)->where('user_name', '=', $input['user_name'])->get();
+                if( is_null($user) || ( count($user)==0 ) ){
+                    return response()->json([
+                            'message' => 'user not found',
+                            'error_code' => ErrorCodeUtility::USER_NOT_FOUND,
+                            'status_code' => '404'
+                        ], 404);
+                }
                 $user[0]->password = bcrypt($input['password']);
                 $user[0]->login_count = 0;
                 $user[0]->save();
                 
+                $verification[0]->delete();
+
                 return $this->response->created();
             }else{
                 return response()->json([
